requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.109.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
packaging==25.0
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.21
starlette==0.35.1
typing_extensions==4.15.0
uvicorn==0.27.0
